Remove unused variables, move auth functions to auth.go
