+++ common/util_test.go
@@ -14,11 +14,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-var (
-	testPrivateKey = []byte("very_sekrit_key")
-	testToken      = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30.AqFWnFeY9B8jj7-l3z0a9iaZdwIca7xhUF3fuaJjU90"
-)
-
 func TestCheckForDockerCompose(t *testing.T) {
 	cwd, err := os.Getwd()
 	assert.Nil(t, err)
