+++ daemon/handler.go
@@ -7,9 +7,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"strings"
 
-	jwt "github.com/dgrijalva/jwt-go"
 	"github.com/docker/docker/api/types"
 	docker "github.com/docker/docker/client"
 	"github.com/google/go-github/github"
@@ -353,39 +351,5 @@ func logHandler(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "text/html")
 		w.WriteHeader(http.StatusOK)
 		fmt.Fprint(w, buf.String())
-	}
-}
-
-// authorized is a function decorator for authorizing RESTful
-// daemon requests. It wraps handler functions and ensures the
-// request is authorized. Returns a function
-func authorized(handler http.HandlerFunc, keyLookup func(*jwt.Token) (interface{}, error)) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		// Collect the token from the header.
-		bearerString := r.Header.Get("Authorization")
-
-		// Split out the actual token from the header.
-		splitToken := strings.Split(bearerString, "Bearer ")
-		if len(splitToken) < 2 {
-			http.Error(w, malformedAuthStringErrorMsg, http.StatusForbidden)
-			return
-		}
-		tokenString := splitToken[1]
-
-		// Parse takes the token string and a function for looking up the key.
-		token, err := jwt.Parse(tokenString, keyLookup)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusForbidden)
-			return
-		}
-
-		// Verify the claims (none for now) and token.
-		if _, ok := token.Claims.(jwt.MapClaims); !ok || !token.Valid {
-			http.Error(w, tokenInvalidErrorMsg, http.StatusForbidden)
-			return
-		}
-
-		// We're authorized, run the handler.
-		handler(w, r)
 	}
 }
